- Move method covariance_to_correlation_matrix to math - fix Dockerfile
