+++ Dockerfile
@@ -17,7 +17,7 @@ RUN pip install .[docs]
 
 # WORKAROUNDS
 ## Qiskit problem in parsing version number
-RUN echo 0.21.0 > /usr/local/lib/python3.8/site-packages/qiskit_terra-0.21.0rc1-py3.8-linux-x86_64.egg/qiskit/VERSION.txt
+RUN echo 0.21.1 > /usr/local/lib/python3.8/site-packages/qiskit_terra-0.21.1-py3.8-linux-x86_64.egg/qiskit/VERSION.txt
 ## Creating folders for mne data
 RUN mkdir /root/mne_data
 RUN mkdir /home/mne_data
