+++ pyriemann_qiskit/utils/__init__.py
@@ -1,4 +1,5 @@
 from . import hyper_params_factory, filtering
+from .math import cov_to_corr_matrix
 from .docplex import (square_cont_mat_var,
                       square_int_mat_var,
                       square_bin_mat_var,
@@ -8,6 +9,7 @@ from .docplex import (square_cont_mat_var,
 __all__ = [
     'hyper_params_factory',
     'filtering',
+    'cov_to_corr_matrix',
     'square_cont_mat_var',
     'square_int_mat_var',
     'square_bin_mat_var',
